backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.35.2
accelerate==0.24.1
huggingface-hub==0.17.3
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
safetensors==0.4.0
opencv-python>=4.8.0
urllib3==1.26.18 